Move the definition of setting classes nearer to their actual use
An intermediate step towards organising packages.
