+++ src/main/scala/com/gu/kindlegen/Article.scala
@@ -5,6 +5,7 @@ import java.time.OffsetDateTime
 import com.gu.contentapi.client.model.v1._
 import com.gu.contentapi.client.utils.CapiModelEnrichment._
 import com.gu.io.{Link, Linkable}
+import com.gu.kindlegen.capi.GuardianProviderSettings
 
 
 case class Article(
