+++ src/main/scala/com/gu/kindlegen/KindleGeneratorSettings.scala
@@ -1,0 +1,22 @@
+package com.gu.kindlegen
+
+import java.nio.file.Path
+
+
+final case class BookBindingSettings(mainSections: Seq[MainSectionTemplate])
+
+
+final case class PublishingSettings(minArticlesPerEdition: Int,
+                                    downloadImages: Boolean,
+                                    prettifyXml: Boolean,
+                                    publicationName: String,
+                                    publicationLink: String,
+                                    files: PublishedFileSettings)
+
+
+final case class PublishedFileSettings(outputDir: Path,
+                                       nitfExtension: String,
+                                       rssExtension: String,
+                                       rootManifestFileName: String) {
+  def encoding = "UTF-8"
+}
