+++ src/main/scala/com/gu/kindlegen/Section.scala
@@ -7,16 +7,6 @@ import com.gu.io.Link.AbsoluteURL
 
 case class Section(id: String, title: String, link: Link)
 
-case class MainSectionTemplate(id: String,
-                               title: Option[String] = None,
-                               link: Option[Link] = None,
-                               overrides: Seq[String] = Nil) {
-  val sectionIds = id +: overrides
-
-  def withDefaults(title: String, link: Link) =
-    Section(id, this.title.getOrElse(title), this.link.getOrElse(link))
-}
-
 object Section {
   def from(tag: Tag): Section = Section(tag.id, tag.webTitle, AbsoluteURL.from(tag.webUrl))
 }
