+++ src/test/scala/com/gu/kindlegen/SettingsSpec.scala
@@ -15,7 +15,9 @@ import org.scalatest.Matchers._
 import com.gu.config.ConfigReader
 import com.gu.contentapi.client.model.v1.TagType
 import com.gu.io.Link.AbsoluteURL
+import com.gu.kindlegen.capi.{ContentApiCredentials, GuardianProviderSettings}
 import com.gu.kindlegen.weather.{WeatherArticleSettings, WeatherSettings}
+import com.gu.kindlegen.weather.accuweather.AccuWeatherSettings
 import com.gu.scalatest.PathMatchers._
